Final fixes/changes and homepage features
