+++ index.html
@@ -2,6 +2,7 @@
 <html>
   <head>
     <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.6.0/jquery.min.js"></script>
+    <script type="module" src="javascript/index.js"></script>
     <link rel="stylesheet" href="styles/index.css"/>
     <title>Visualization Tool</title>
   </head>
@@ -13,15 +14,29 @@
     </header>
     <div class="pageContainer">
       <div id="introduction">
-        <h1 class="titleSecondary">
+        <h1>
           The All-Purpose Visualization and Solver Tool
         </h1>
-        <p class="subtitleSecondary">
+        <p>
           On Every Computer Science Topic I've Learned
         </p>
       </div>
       <div id="tool-list">
-
+        <div id="pathfinder-tool">
+          <img src="assets/pathfinder-feature.png"/>
+          <h1>Grid Pathfinder Tool</h1>
+          <p>
+            Pathfinding is about having an algorithm find a route between two points. It can find the shortest path, and can also work with paths with costs.
+          </p>
+          <p>Learn more about such algorithms by finding paths on a grid here!</p>
+        </div>
+        <div id="sorting-tool">
+          <img src="assets/coming-soon.jpg"/>
+          <h1>Sorting Tool</h1>
+          <p>
+            Coming soon!
+          </p>
+        </div>
       </div>
     </div>
   </body>
