+++ javascript/index.js
@@ -1,0 +1,7 @@
+$(document).ready(function() {
+    $(".toolFeatureBox").click(function() {
+        const fileName = $(this).attr("id").split("-")[0];
+        if(fileName !== "sorting")
+            window.location.href += fileName;
+    });
+});
