+++ navbar.html
@@ -5,21 +5,22 @@
   <body>
     <div id="navbar">
       <a id="visualization-tool-home" href="/">
-        <h1 class="subtitlePrimary">Visualization Tool</h1>
+        <h1>Visualization Tool</h1>
       </a>
       <a href="/">
         <button class="navbarButton">
-          <p class="subtitlePrimary">Home</p>
+          <p>Home</p>
         </button>
       </a>
-      <a href="graph_theory">
+      <a href="pathfinder">
         <button class="navbarButton">
-          <p class="subtitlePrimary">Graph & Pathfinder</p>
+          <p>Grid Pathfinder</p>
         </button>
       </a>
-      <a href="sorting">
+      <a href="javascript:void(0)" disabled="true">
         <button class="navbarButton">
-          <p class="subtitlePrimary">Sorting</p>
+          <p>Sorting</p>
+          <p>(Coming soon!)</p>
         </button>
       </a>
     </div>
