+++ styles/index.css
@@ -1,43 +1,6 @@
 .typography {
   font-family: "Helvetica Neue", "Helvetica", "Arial", sans-serif;
 }
-.unselectable {
-  user-select: none;
-  -moz-user-select: none;
-  -khtml-user-select: none;
-  -webkit-user-select: none;
-  -o-user-select: none;
-}
-.titlePrimary {
-  font-size: 48px;
-  font-family: "Helvetica Neue", "Helvetica", "Arial", sans-serif;
-  color: black;
-}
-.titleSecondary {
-  font-size: 48px;
-  font-family: "Helvetica Neue", "Helvetica", "Arial", sans-serif;
-  color: white;
-}
-.subtitlePrimary {
-  font-size: 36px;
-  font-family: "Helvetica Neue", "Helvetica", "Arial", sans-serif;
-  color: black;
-}
-.subtitleSecondary {
-  font-size: 36px;
-  font-family: "Helvetica Neue", "Helvetica", "Arial", sans-serif;
-  color: white;
-}
-.textPrimary {
-  font-size: 18px;
-  font-family: "Helvetica Neue", "Helvetica", "Arial", sans-serif;
-  color: black;
-}
-.textSecondary {
-  font-size: 18px;
-  font-family: "Helvetica Neue", "Helvetica", "Arial", sans-serif;
-  color: white;
-}
 html {
   height: 100%;
   box-sizing: border-box;
@@ -57,18 +20,178 @@ body {
   padding-top: 65px;
   box-sizing: border-box;
 }
+.modal {
+  position: fixed;
+  left: 0;
+  top: 0;
+  width: 100%;
+  height: 100%;
+  opacity: 0;
+  visibility: hidden;
+  transform: scale(1.1);
+}
+.modal .modalContent {
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  width: 50%;
+  height: 50%;
+  transform: translate(-50%, -50%);
+  background-color: white;
+  padding: 16px;
+  border-radius: 10px;
+}
+.modal .closeButton {
+  float: right;
+  margin: 0 6px 0 auto;
+  cursor: pointer;
+}
+.modal.showModal {
+  opacity: 1;
+  visibility: visible;
+  transform: scale(1);
+  transition: visibility 0s linear 0s, opacity 0.25s 0s, transform 0.25s;
+  background-color: rgba(0, 0, 0, 0.75);
+  z-index: 10;
+}
+.unselectable {
+  user-select: none;
+  -moz-user-select: none;
+  -khtml-user-select: none;
+  -webkit-user-select: none;
+  -o-user-select: none;
+}
+.hidden {
+  display: none;
+}
+.errorMessage {
+  font-size: 12px !important;
+  color: #ee4b2b !important;
+}
+@keyframes horizontal-shaking {
+  0% {
+    transform: translateX(0);
+  }
+  25% {
+    transform: translateX(15px);
+  }
+  50% {
+    transform: translateX(-15px);
+  }
+  75% {
+    transform: translateX(15px);
+  }
+  100% {
+    transform: translateX(0);
+  }
+}
+.shakingAnimation {
+  animation-name: horizontal-shaking;
+  animation-duration: 0.2s;
+  animation-fill-mode: forwards;
+}
 #introduction {
   width: 100%;
   min-height: 250px;
   max-height: 500px;
   height: 40%;
   display: flex;
-  background-color: #30D5C8;
+  background-color: #0e9594;
   flex-direction: column;
   justify-content: flex-start;
   align-items: center;
   text-align: center;
 }
 #introduction h1 {
+  font-size: 48px;
+  font-family: "Helvetica Neue", "Helvetica", "Arial", sans-serif;
+  color: white;
   margin-top: 8%;
+}
+#introduction p {
+  font-size: 36px;
+  font-family: "Helvetica Neue", "Helvetica", "Arial", sans-serif;
+  color: white;
+  padding-bottom: 12px;
+  font-size: 24px;
+}
+#tool-list {
+  display: flex;
+  flex-direction: row;
+  flex-wrap: wrap;
+  justify-content: space-evenly;
+}
+#tool-list #pathfinder-tool {
+  width: 440px;
+  height: 520px;
+  margin: 20px;
+  border-radius: 10px;
+  border: 2px solid black;
+  background-color: #f1fffa;
+  display: flex;
+  flex-direction: column;
+  -webkit-transition: -webkit-transform 0.2s;
+  transition: -webkit-transform 0.2s;
+  transition: transform 0.2s;
+  transition: transform 0.2s, -webkit-transform 0.2s;
+}
+#tool-list #pathfinder-tool img {
+  width: 100%;
+  border-top-left-radius: 10px;
+  border-top-right-radius: 10px;
+  border-bottom: 2px solid black;
+}
+#tool-list #pathfinder-tool h1 {
+  font-size: 36px;
+  font-family: "Helvetica Neue", "Helvetica", "Arial", sans-serif;
+  color: #0e9594;
+  font-size: 28px;
+  margin: 3px 6px 12px 6px;
+}
+#tool-list #pathfinder-tool p {
+  font-size: 18px;
+  font-family: "Helvetica Neue", "Helvetica", "Arial", sans-serif;
+  color: #0e9594;
+  margin: 0 6px;
+}
+#tool-list #pathfinder-tool:hover {
+  -webkit-transform: scale(1.05);
+  transform: scale(1.05);
+}
+#tool-list #sorting-tool {
+  width: 440px;
+  height: 520px;
+  margin: 20px;
+  border-radius: 10px;
+  border: 2px solid black;
+  background-color: #FCE1C5;
+  display: flex;
+  flex-direction: column;
+  -webkit-transition: -webkit-transform 0.2s;
+  transition: -webkit-transform 0.2s;
+  transition: transform 0.2s;
+  transition: transform 0.2s, -webkit-transform 0.2s;
+}
+#tool-list #sorting-tool img {
+  width: 100%;
+  border-top-left-radius: 10px;
+  border-top-right-radius: 10px;
+  border-bottom: 2px solid black;
+}
+#tool-list #sorting-tool h1 {
+  font-size: 36px;
+  font-family: "Helvetica Neue", "Helvetica", "Arial", sans-serif;
+  color: #cc4bc2;
+  font-size: 28px;
+  margin: 3px 6px 12px 6px;
+}
+#tool-list #sorting-tool p {
+  font-size: 18px;
+  font-family: "Helvetica Neue", "Helvetica", "Arial", sans-serif;
+  color: #cc4bc2;
+  margin: 0 6px;
+}
+#tool-list #sorting-tool:hover {
+  -webkit-transform: scale(1.05);
+  transform: scale(1.05);
 }
