+++ styles/navbar.css
@@ -1,43 +1,6 @@
 .typography {
   font-family: "Helvetica Neue", "Helvetica", "Arial", sans-serif;
 }
-.unselectable {
-  user-select: none;
-  -moz-user-select: none;
-  -khtml-user-select: none;
-  -webkit-user-select: none;
-  -o-user-select: none;
-}
-.titlePrimary {
-  font-size: 48px;
-  font-family: "Helvetica Neue", "Helvetica", "Arial", sans-serif;
-  color: black;
-}
-.titleSecondary {
-  font-size: 48px;
-  font-family: "Helvetica Neue", "Helvetica", "Arial", sans-serif;
-  color: white;
-}
-.subtitlePrimary {
-  font-size: 36px;
-  font-family: "Helvetica Neue", "Helvetica", "Arial", sans-serif;
-  color: black;
-}
-.subtitleSecondary {
-  font-size: 36px;
-  font-family: "Helvetica Neue", "Helvetica", "Arial", sans-serif;
-  color: white;
-}
-.textPrimary {
-  font-size: 18px;
-  font-family: "Helvetica Neue", "Helvetica", "Arial", sans-serif;
-  color: black;
-}
-.textSecondary {
-  font-size: 18px;
-  font-family: "Helvetica Neue", "Helvetica", "Arial", sans-serif;
-  color: white;
-}
 html {
   height: 100%;
   box-sizing: border-box;
@@ -56,6 +19,76 @@ body {
   height: 100%;
   padding-top: 65px;
   box-sizing: border-box;
+}
+.modal {
+  position: fixed;
+  left: 0;
+  top: 0;
+  width: 100%;
+  height: 100%;
+  opacity: 0;
+  visibility: hidden;
+  transform: scale(1.1);
+}
+.modal .modalContent {
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  width: 50%;
+  height: 50%;
+  transform: translate(-50%, -50%);
+  background-color: white;
+  padding: 16px;
+  border-radius: 10px;
+}
+.modal .closeButton {
+  float: right;
+  margin: 0 6px 0 auto;
+  cursor: pointer;
+}
+.modal.showModal {
+  opacity: 1;
+  visibility: visible;
+  transform: scale(1);
+  transition: visibility 0s linear 0s, opacity 0.25s 0s, transform 0.25s;
+  background-color: rgba(0, 0, 0, 0.75);
+  z-index: 10;
+}
+.unselectable {
+  user-select: none;
+  -moz-user-select: none;
+  -khtml-user-select: none;
+  -webkit-user-select: none;
+  -o-user-select: none;
+}
+.hidden {
+  display: none;
+}
+.errorMessage {
+  font-size: 12px !important;
+  color: #ee4b2b !important;
+}
+@keyframes horizontal-shaking {
+  0% {
+    transform: translateX(0);
+  }
+  25% {
+    transform: translateX(15px);
+  }
+  50% {
+    transform: translateX(-15px);
+  }
+  75% {
+    transform: translateX(15px);
+  }
+  100% {
+    transform: translateX(0);
+  }
+}
+.shakingAnimation {
+  animation-name: horizontal-shaking;
+  animation-duration: 0.2s;
+  animation-fill-mode: forwards;
 }
 a {
   text-decoration: none;
@@ -77,11 +110,16 @@ a {
 #navbar-location #navbar #visualization-tool-home {
   margin: 0 auto 0 15px;
 }
+#navbar-location #navbar #visualization-tool-home h1 {
+  font-size: 36px;
+  font-family: "Helvetica Neue", "Helvetica", "Arial", sans-serif;
+  color: black;
+}
 #navbar-location #navbar #visualization-tool-home:hover {
   cursor: pointer;
 }
 #navbar-location #navbar .navbarButton {
-  width: 120px;
+  width: 180px;
   height: 100%;
   text-align: center;
   background-color: white;
@@ -89,8 +127,11 @@ a {
 }
 #navbar-location #navbar .navbarButton p {
   font-size: 18px;
+  font-family: "Helvetica Neue", "Helvetica", "Arial", sans-serif;
+  color: black;
   margin: auto;
 }
 #navbar-location #navbar .navbarButton:hover {
+  background-color: #bbc4cc;
   cursor: pointer;
 }
